fixing minor bug and made changes to forumreview (final)

+++ lib/screens/ordernborrow screens/client/borrow/books.dart
@@ -344,7 +344,7 @@ class _BooksPageState extends State<BooksPage> {
                         )),
                     subtitle: Text(snapshot.data![index].fields.kategori),
                     trailing: RatingBar.builder(
-                      initialRating: snapshot.data![index].fields.rating,
+                      initialRating: snapshot.data![index].fields.rating.toDouble(),
                       minRating: 0,
                       direction: Axis.horizontal,
                       itemCount: 5,
