+++ lib/screens/ordernborrow screens/client/borrow/borrowedbooks.dart
@@ -338,7 +338,7 @@ class _BorrowedBooksState extends State<BorrowedBooks> {
                         )),
                     subtitle: Text(snapshot.data![index].fields.kategori),
                     trailing: RatingBar.builder(
-                      initialRating: snapshot.data![index].fields.rating,
+                      initialRating: snapshot.data![index].fields.rating.toDouble(),
                       minRating: 0,
                       direction: Axis.horizontal,
                       itemCount: 5,
